fix(pe): correct two review findings in the views and the append seam
Implementation review of Tasks 1-3 found two consequential defects. Both
verified against source, and the second by a deliberate compile probe.

1. PeSparseScratch::vertexInputs was 2, but the V2 schema caps the section at
   maxCount 1 with V2SectionRuleSingle: FVF versus vertex declaration is the
   entry's `kind` field, not two entries, and the legacy shim's own scratch is
   sized 1. A 2-entry span is rejected by validSectionCount, so Task 7's
   producer -- which the wrong comment actively invited to fill both slots --
   would have dropped draws at runtime with D3DERR_INVALIDCALL.

   Rather than fix the one number, every scratch capacity is now
   static_asserted against v2SectionRule(kind)->maxCount, so a schema change
   breaks the build here instead of dropping draws. PeBindingView::textures is
   likewise tied to kPeTextureSlots, which it previously matched only by
   coincidence. The header's "capacities match the section caps" claim is now
   enforced rather than asserted in prose.

2. appendRecordV2 documented an HRESULT emitter but did not constrain it.
   HRESULT is LONG, so an emitter written the way the plan's own Task 5/6
   snippets showed -- a lambda with no trailing return type wrapping a
   bool-returning V2 emitter -- deduces bool and converts false to 0 == S_OK:
   a failed append reported as success, CapacityPost run, and
   notePeChunkAppendBoundary counting a record that never landed. Added a
   static_assert on the invoke result.

   My first attempt to prove the assert worked was invalid: the legacy
   adapter's lambda declares `-> HRESULT`, so bool converted inside the lambda
   and decltype still saw HRESULT. Re-probed with a deduced-bool emitter, which
   fails as intended ("is_same_v<bool, long>").

Also: dropped the no-op casts Task 1 Step 3 called for and left behind, and
replaced line-number citations in the new headers with symbol names -- this
migration moves large blocks of d3d9_pe_device.cpp and the numbers had already
drifted by ~66 lines. The plan's Task 5-9 emitter snippets and vertexInputs
capacity are corrected in the same commit so the remaining tasks are read
against the fixed text.

Verified: 669/669 meson tests; all four build lanes.

+++ src/d3d9/d3d9_pe_state_shadow.hpp
@@ -329,28 +329,25 @@ inline bool vertexTextureSamplerSlot(std::uint32_t sampler, std::uint32_t& slot)
     if (sampler < kD3dVertexTextureSampler0 || sampler > kD3dVertexTextureSampler3) {
         return false;
     }
-    slot = kPeFragmentSamplerSlots +
-        static_cast<std::uint32_t>(sampler - kD3dVertexTextureSampler0);
+    slot = kPeFragmentSamplerSlots + (sampler - kD3dVertexTextureSampler0);
     return true;
 }
 
 inline bool textureBindingSlot(std::uint32_t stage, std::uint32_t& slot) noexcept {
     if (stage < kPeFragmentSamplerSlots) {
-        slot = static_cast<std::uint32_t>(stage);
+        slot = stage;
         return true;
     }
     return vertexTextureSamplerSlot(stage, slot);
 }
 
-inline std::uint32_t textureStageStateSlot(
-    std::uint32_t type) noexcept {
-    return std::min<std::uint32_t>(
-        static_cast<std::uint32_t>(type), kPeTextureStageStateSlots - 1u);
+inline std::uint32_t textureStageStateSlot(std::uint32_t type) noexcept {
+    return std::min<std::uint32_t>(type, kPeTextureStageStateSlots - 1u);
 }
 
 inline bool samplerSlot(std::uint32_t sampler, std::uint32_t& slot) noexcept {
     if (sampler < kPeFragmentSamplerSlots) {
-        slot = static_cast<std::uint32_t>(sampler);
+        slot = sampler;
         return true;
     }
     return vertexTextureSamplerSlot(sampler, slot);
@@ -358,7 +355,7 @@ inline bool samplerSlot(std::uint32_t sampler, std::uint32_t& slot) noexcept {
 
 inline bool samplerStateSlot(std::uint32_t type,
                              std::uint32_t& slot) noexcept {
-    slot = static_cast<std::uint32_t>(type);
+    slot = type;
     return slot < kPeSamplerStateSlots;
 }
